+++ tests/legacy/system/test_edb_padstacks.py
@@ -23,6 +23,7 @@
 """Tests related to Edb padstacks
 """
 import os
+from pathlib import Path
 
 import pytest
 
@@ -458,3 +459,16 @@ class TestClass:
         assert "main_via" in edbapp.padstacks.definitions
         assert "via_central" in edbapp.padstacks.definitions
         edbapp.close()
+
+    def test_reduce_via_in_bounding_box(self):
+        source_path = Path(__file__).parent.parent.parent / "example_models" / "TEDB" / "vias_300.aedb"
+        edbapp = Edb(edbpath=source_path)
+        assert len(edbapp.padstacks.instances) == 301
+        # empty bounding box
+        assert edbapp.padstacks.reduce_via_in_bounding_box([-16e-3, -7e-3, -13e-3, -6e-3], 10, 10) is False
+        # over sampling
+        assert edbapp.padstacks.reduce_via_in_bounding_box([-20e-3, -10e-3, 20e-3, 10e-3], 20, 20) is False
+
+        assert edbapp.padstacks.reduce_via_in_bounding_box([-20e-3, -10e-3, 20e-3, 10e-3], 10, 10) is True
+        assert len(edbapp.padstacks.instances) == 96
+        edbapp.close_edb()
